TS - 8 Creating BarElement component
- Creating BarElement component to enable targeting the bar element with useRef;
- Implementing the BarElement component in AddEvent component and Bar component;
- Creating a new folder for Bar components.
- Adding other changes from the pull request review
