+++ src/pages/HomePage.tsx
@@ -7,7 +7,7 @@ import { Button, Row, Col, Form } from "react-bootstrap";
 import useFireStore from "../hooks/useFireStore";
 import { useEffect, useState, useContext } from "react";
 import EventList from "../components/EventList";
-import Bar from "../components/Bar";
+import Bar from "../components/Bar/Bar";
 import LoadingBar from "../components/LoadingBar";
 import pieChartIcon from "../assets/pie-chart-fill.svg";
 import { toast } from "react-toastify";
@@ -15,8 +15,7 @@ import DateInput from "../components/DateInput";
 import { dataReformer } from "../utils/reformDataForBar";
 
 const HomePage = () => {
-  const { date, data, setIsLoading, onLogout, setData, setAdding, loading } =
-    useContext(Context);
+  const { date, data, setIsLoading, onLogout, setData, setAdding, loading } = useContext(Context);
   const firestore = useFireStore();
   const [isPieChart, setIsPieChart] = useState(false);
   const filteredData = data.filter((elem) => elem.date === date);
@@ -58,9 +57,7 @@ const HomePage = () => {
           {isPieChart && !!filteredData.length && <Pie />}
           <Bar />
           {!filteredData[0]?.event?.length ? (
-            <p style={{ textAlign: "center" }}>
-              You haven`t added any events for this day yet
-            </p>
+            <p style={{ textAlign: "center" }}>You haven`t added any events for this day yet</p>
           ) : (
             <EventList />
           )}
@@ -72,11 +69,7 @@ const HomePage = () => {
                 </Button>
               </Col>
               <Col xl={4} m={4} sm={4} xs={12}>
-                <Button
-                  className="mt-3"
-                  onClick={addNewTaskButtonHandler}
-                  variant="secondary"
-                >
+                <Button className="mt-3" onClick={addNewTaskButtonHandler} variant="secondary">
                   Add new task
                 </Button>
               </Col>
@@ -88,11 +81,7 @@ const HomePage = () => {
                     setIsPieChart((prev) => !prev);
                   }}
                 >
-                  <img
-                    style={{ marginRight: "0.5rem" }}
-                    src={pieChartIcon}
-                    alt=""
-                  />
+                  <img style={{ marginRight: "0.5rem" }} src={pieChartIcon} alt="" />
                   Pie chart
                 </Button>
               </Col>
